requirements.txt
opencv-python
numpy
pyserial
pyttsx3
tensorflow
tflite-runtime
openai
pydantic
RPi.GPIO
